define config file, to make import cleaner
