+++ error-response-core/src/main/java/com/itachallenge/errorcore/exceptionhandler/BaseExceptionHandler.java
@@ -13,14 +13,12 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 import org.springframework.web.server.ResponseStatusException;
 import org.springframework.web.server.ServerWebInputException;
 
 @Slf4j
 @RequiredArgsConstructor
-@RestControllerAdvice
 public abstract class BaseExceptionHandler {
 
     protected final ErrorResponseBuilder responseBuilder;
